requirements.txt
apache-airflow-providers-snowflake==2.3.1
snowflake-connector-python==2.7.1
snowflake-sqlalchemy==1.2.4
pytz==2021.3